+++ lua/obsidian/init.lua
@@ -107,11 +107,6 @@ obsidian.setup = function(opts)
     cmp.register_source("obsidian_tags", require("cmp_obsidian_tags").new())
   end
 
-  -- Setup UI add-ons.
-  if client.opts.ui.enable then
-    obsidian.ui.setup(client.opts.ui)
-  end
-
   local group = vim.api.nvim_create_augroup("obsidian_setup", { clear = true })
 
   -- Complete setup and update workspace (if needed) when entering a markdown buffer.
@@ -135,6 +130,7 @@ obsidian.setup = function(opts)
       if not client.current_workspace.locked and workspace ~= client.current_workspace then
         log.debug("Switching to workspace '%s' @ '%s'", workspace.name, workspace.path)
         client:set_workspace(workspace)
+        client:update_ui(ev.buf)
       end
 
       -- Register mappings.
@@ -185,13 +181,7 @@ obsidian.setup = function(opts)
         return
       end
 
-      local frontmatter = nil
-      if client.opts.note_frontmatter_func ~= nil then
-        frontmatter = client.opts.note_frontmatter_func(note)
-      end
-
-      local updated = note:save_to_buffer(bufnr, frontmatter)
-      if not client._quiet and updated then
+      if client:update_frontmatter(note, bufnr) then
         log.info "Updated frontmatter"
       end
     end,